requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
Django==6.0.3
djangorestframework==3.17.1
drf-yasg==1.21.15
idna==3.11
inflection==0.5.1
packaging==26.0
pillow==12.1.1
python-decouple==3.8
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.33.0
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
duffel-api==0.6.2
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
python-dateutil==2.9.0
